mcp-server-01/requirements.txt
mcp[cli]
python-dotenv
httpx